adopt SDK for new config
