+++ sdk/src/backend/index.ts
@@ -1,3 +1,4 @@
 export * from "./getReceipt";
 export * from "./getSendPayload";
 export * from "./getTokensConfig";
+export type * from "./types";
