+++ sdk/src/backend/types/index.d.ts
@@ -1,0 +1,3 @@
+export type * from "./receipt";
+export type * from "./payload";
+export type * from "./tokenConfig";
