+++ sdk/tsconfig.json
@@ -13,7 +13,7 @@
     "allowJs": true,
     "outDir": "dist",
     "rootDir": "./src",
-    "typeRoots": ["./node_modules/@types"]
+    "typeRoots": ["./node_modules/@types", "./src/backend/types", "./src/evm/types"],
   },
   "include": ["src", "src/**/*.json"]
 }
